ecommerce_backend/requirements.txt
asgiref==3.8.1
bcrypt==3.2.0
certifi==2024.12.14
cryptography==3.4.8
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.17.0
elasticsearch==8.17.0
elasticsearch-dsl==8.17.1
future==0.18.2
Pillow==9.0.1
protobuf==3.12.4
psycopg2==2.9.10
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2022.1
redis==5.2.1
selenium==4.27.1
sqlparse==0.5.3
urllib3==1.26.5
websocket-client==1.8.0
gunicorn==20.1.0
